requirements.txt
webdriver-manager==3.5.4
selenium==4.1.0
pytest==6.0.2
pytest-html==2.1.1
requests==2.27.1
pip==21.1.3
jsonpath==0.82
allure-pytest==2.9.43
allure-python-commons==2.9.43
virtualenv==2.10.0